backend/Url-Analyzer/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pyOpenSSL==23.3.0
urllib3==2.1.0
gunicorn==21.2.0
